Don't reuse invalidated distance field cache
When the fonts are invalidated e.g. because application fonts are
cleared, there will be a bit of a round-trip where we reset all
text components to rematch the fonts they use.

Since the scenegraph nodes contain direct pointers to the distance
field cache, we cannot delete the cache itself before the text has
been reset and a new font has been matched.

This caused an issue where an application font was unloaded and
a new one was loaded instead. Since they both had the same
virtual file name (":qmemoryfonts/0"), however, we would get
the glyph cache for the old font when we were resetting the text
with the old. Since the glyph indexes referred to the new font,
we would end up with random glyphs in the text.

The core problem was that the stale caches were still in the
hash when the scenegraph was being reset. Instead, we start
the sync by either deleting inactive caches or moving the
active ones into a separate list. When the text items are
reset, the hash is populated with new caches. When the sync
is done, we flush the stale caches: Meaning that we delete
the ones that are now no longer used.

Pick-to: 6.10
Task-number: QTBUG-132108
Change-Id: Iae4bebdbd1acc34ff86a2f9f6f34fdd0d6cbac0e
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>