ColorPicker: remove handle as a listener during destruction
Children get destroyed before their parents, and objects notify
listeners when they get destroyed. If a child is a listener on a parent,
then this results in a (partially) destroyed listener object being
notified, resulting in UB and - fortunately - test-breaking warnings.

The handle is (typically) a child of the dialog, and added as a listener
for implicit size changes when set. Explicitly remove it in the
destructor of QQuickAbstractColorPicker, before destruction of the
object tree.

Pick-to: 6.10 6.9 6.8 6.5
Fixes: QTBUG-137561
Change-Id: I76ff1ff8bc7d563133d7bb9cd8b7bee76d6cec9b
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>