Mark src/qmlnetwork as security-significant
The classes in this subdirectory simply expose some of the QtNetwork
classes to QML, and mostly act like proxies.
The only interesting cases are related to reading ssl certificates
from files, but in this case we just take the user-provided file path
and use it in QFile without any preprocessing. So it's QFile's
responsibility to parse the path.

Fixes: QTBUG-136204
Pick-to: 6.10 6.9 6.8
Change-Id: Ia62fb8f7c8539c6461236b4c519a9371d7a03871
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>