Software Adaptation: Disable partial updates when fractional scaling
When using a scaling factor for the UI that is not an integer, the
partial update mechanism in the software Adapation can break down and
lead to many artifacts. Now we check if we're trying to render with
factional scaling and disable partial updates. It's possible to opt-out
of this by setting the environment variable:

QSG_SOFTWARE_RENDERER_FORCE_PARTIAL_UPDATES=1

It is also possible to disable partial updates using the same variable

QSG_SOFTWARE_RENDERER_FORCE_PARTIAL_UPDATES=0

Fixes: QTBUG-136943
Pick-to: 6.10 6.8
Change-Id: If4b0bc403a96ab037f213e89708371a49bc8ae02
Reviewed-by: Christian Strømme <christian.stromme@qt.io>