Allow triangle fan in QSGGeometry when the underlying API supports it
Will not work with D3D and Metal. With those we will continue to reject
QSGGeometries having a drawing mode of DrawTrangineFan with a warning.

Relevant mainly for OpenGL and applications ported from Qt 5. For a
cross-platform, portable solution they would need to move away from
triangle fans, but in the meantime, if the intent is to function just
with OpenGL on Qt 6, it is good for compatibility if we enable this.

Pick-to: 6.10 6.9 6.8
Change-Id: Iaa9376234ef854557458e5b6020e39c60666ad91
Reviewed-by: Andy Nichols <andy.nichols@qt.io>