Doc: Describe when Repeater.count doesn't match delegate instance count
Clarifies the known conditions when Repeater.count does not match the
actual number of delegate instances: if it doesn't have a visual parent
("set up correctly" was a bit vague, raising the question of whether
there are other forms of incorrectness), or because it's temporarily out
of sync during delegate instantiation/deletion. If you need more info,
monitor the signals and/or check itemAt().

Amends a6eac7e299016c9ee8742abffd780683adac3066

Pick-to: 6.10 6.9 6.8
Fixes: QTBUG-127955
Change-Id: I6862beb8d61b1cf8a747bdc3ad02226691e6c3eb
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>