Controls: speed up animations in auto tests
We do this by first introducing a helper:
QQuickVisualTestUtils::setFastAnimations. This does the following:

    QUnifiedTimer::instance()->setSpeedModifier(fastAnimations ? 5 : 1);

Tests with longer execution times now call this.

However, for the QML-only tests in tests/auto/quickcontrols/controls,
we need to make some extra adjustments, as setSpeedModifier must be
called after the QCoreApplication has been created (because
QUnifiedTimer::instance connects to QCoreApplication::aboutToQuit).
It's impossible to do that when using quick_test_main, so we switch to
quick_test_main_with_setup, which provides the applicationAvailable
callback.

Fixes: QTBUG-137919
Change-Id: I08beb9281777b96f4ac747005e2cdd4b4ecf045c
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>