QQmlCodeModelManager: add method to find project root
Add a method that returns the project root for a file. The project root
is the workspace with the shortest url that contains some file.

This will be needed when searching for C++ headers, as they might be in
a different subproject/workspace than the currently edited file.

Task-number: QTBUG-128393
Change-Id: I63ce3895abb20c9851dc77da3b4d0c4bd267c871
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>