Remove QML generator from SVG manual test
The QML generator code path in the test has some bugs due to how it
generates embedded images. In addition, it does not map to loading
other file types such as Lottie in the same application, since
that would need special handling.

Since it is now tested through the item generator (the item
generator uses it internally), we don't need a separate UI for
this in the test.

Task-number: QTBUG-135267
Pick-to: 6.10
Change-Id: I9e35bccebcd9e763c74ce44cf9d4e89fc0d1a802
Reviewed-by: Hatem ElKharashy <hatem.elkharashy@qt.io>
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>