tst_qmlcachegen: unify behavior for cross compiled tests
Create a new macro that skips the test, and use it consistently around
the tests, instead of sometimes skipping and sometimes failing the
tests.

Also add a macro that prints the same message and returns false for the
generateCpp and generateCache helpers, to help in the case someone
forgets to add the skipping macro to their new test.

Change-Id: I91d35a32f345df1ebb0616046923a8c9b6b515e6
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>