LoggingCategory: Ensure that name is set early enough
Every LoggingCategory needs to have a name. However, it might be used
before QQmlParserStatus::componentComplete has run. In theory, it might
even be used before the binding has been installed.

To fix this, force the completion logic to run early if the category is
used.
Given that the name of the logging category must never change, we make
the assumption that no complex bindings are used for name, but only
literal strings. Consequently, we don't introduce the overhead of
flushing bindings, which would be necessary if we were to support them.

Fixes: QTBUG-116539
Pick-to: 6.10
Change-Id: Ic194a7dcbae25b5f982198e9ea27f806c84e5cac
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>