Loader: Also drop engine when clearing contexts recursively
Simply clearing the expressions is not enough. We might have internal
signal/slot connections in addition. The main protagonists of such
behavior listen to context invalidation, though, and a context without
engine counts as invalid. At the same time, we leave the context object
in place so that we don't re-open QTBUG-66822. The currently running
binding will still be able to complete this way.

Pick-to: 6.10 6.9 6.8 6.5
Fixes: QTBUG-118188
Change-Id: Ia973be26f0e4e3244adc71f95df7e0b4bef412c3
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>