QQmlTypeLoader: Also populate directory cache on directoryExists
Previously, QQmlTypeLoader had different cache entries for
directoryExists() and fileExist() in the same cache. The ones for
fileExists() had an extra slash appended to their path.

This was entirely pointless because a path is a path and if it exists,
we can use it for file lookup. Furthermore, it was wasteful because it
halved the capacity of the directory cache while adding duplicate keys.
However, it did mask one bug: The entries for directories could stay
unpopulated without disturbing the ones for files.

The masking was not complete, though. When inserting a directory entry
that ended in a double slash, one of the slashes was removed, leaving
the entry in the form expected for files.

Now we always remove all trailing slashes, and always populate the cache
when an entry is first inserted.

Task-number: QTBUG-134652
Change-Id: I333c81655d6f6dfff1e46dae8eef3c46a33d97ac
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>