painterpathquickshape test: Add designhelpers_rectangleshape.qml
Lancelot at some point complained about AA changes in (among others) the
designhelpers_rectangleshape.qml baseline.  By having a closer look, it
became clear that the issue was actually overlapping triangles on arc
triangulation. It's not the only case where those occur, but it seems
like a good test case to have, because rounded rectangles are so common.

This is not the complete set of shapes from
designhelpers/designhelpers_rectangleshape.qml :
just a few that were flagged in Lancelot, plus a ControlledShape
that reproduces a similar shape and is adjustable in the usual ways.

Change-Id: Ic87f51e284157ca37b0e693e057a5ff9bdba3dcf
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>