QtQml: Remove QQml_isFileCaseCorrect()
QQmlTypeLoader::fileExists() can do the same job, on all platforms, and
more efficiently.

Drop the file name case check when loading plugins. Plugin paths are
platform-specific and controlled by the build system. There is no sane
way to confuse the casing. Also, the build system apparently creates
case-mismatched plugins on android, and that works fine so far because
QQml_isFileCaseCorrect() was a no-op on android.

Fixes: QTBUG-134652
Change-Id: I46059be7998ff8402b400b3e2a5385065275e38d
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>