ScriptFormatter: Preserve user spacing around comments
We want to respect the formatting and spacing of the user around
comments. Comments already contain the information about spaces and new
lines before and after them and they are included when writing the
comment.

However, this conflicts with the need the add spaces around tokens in
the non-commented case. We need to only ensure a space if it doesn't
come immediately before or after a comment since the comment already
includes the spacing. Otherwise we would be duplicating the spaces.

In order to do this, keep track of whether the last write was a comment
and defer actually adding the spaces until we know no comment follows.

This fixes errors where extra spaces were inserted around comments.

Tests needed to be updated to adhere to the rule that we respect user
spacing around comments.

The SpaceBeforePostComment option is also no longer needed now and was
removed. The remaining options were also renamed to better reflect their
behavior.

Task-number: QTBUG-133315
Change-Id: I38e7fb65ac76ead6287c9dcb2f43f8fdd09347da
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>