Move a copy of QQuickAttachedPropertyPropagator to Qt Quick
In order to fix QTBUG-134903, we need to introduce a
QQuickContextMenuAttached type so that users have a way to opt out of
text editing context menus. This type lives in src/quicktemplates and
would need to derive from QQuickAttachedPropertyPropagator, but that
lives in src/quickcontrols. This relationship isn't possible; controls
depends on templates, not the other way around.

Instead, introduce a private copy of QQuickAttachedPropertyPropagator
to quick/utils in a Private namespace, so that we can use the same
name for the sake of simplicity when it comes to making this new type
public and removing the copy in Controls.

Text editing controls need access to findAttachedParent (and, by
extension, firstCppMetaObject) in order to find the
QQuickContextMenuAttached attached to it (or one of its parents)
_without_ creating an instance of it. For that reason, we move
QQuickAttachedPropertyPropagatorPrivate out into its own private header
and move all of the static functions in there for consistency.

Task-number: QTBUG-133556
Task-number: QTBUG-134903
Change-Id: I4f8013fb3480ee1177db812dad9b7b1fd646fb07
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>