FluentWinUI3: Improve the ToolButton when using high contrast themes
The AppBar[Toggle]Button is the equivalent WinUI3 control.
Our ToolButton should have the same look and feel as the AppBar*
buttons both with and without enabling a high contrast theme.

Interestingly enough, these AppBar* buttons differs from the regular
buttons in that they look identical when they're pressed/hovered, and
checked. Furthermore, if they're both checked and hovered, they look
like a regular button that is hovered, but not checked.

To reduce the amount of logic in ButtonBackground.qml, we're now allowing
users of that object type to provide their own function for calculating
the background color.

Task-number: QTBUG-129088
Pick-to: 6.11 6.10
Change-Id: I891d5492f36a76ea187ffb3a5d4d03a033501623
Reviewed-by: Doris Verria <doris.verria@qt.io>