FluentWinUI3: Give [Double]SpinBox an outline for high contrast themes
Replace the regular background with a simple rectangle, that has a
border that changes color based on the state of the SpinBox.

Also change the palette role that is used for the text for both
SpinBox and ComboBox to be buttonText. This was previously set to
text, which caused the color to be incorrect when using
the "Night Sky" contrast theme.

The up and down indicators for SpinBox are also adjusted when using
high contrast themes using ColorImage.

Task-number: QTBUG-129088
Pick-to: 6.11 6.10
Change-Id: I4af4a8b3c6f2ad7993bc23a9b73e7d0dc2cfdbbc
Reviewed-by: Doris Verria <doris.verria@qt.io>