Add a static create function to QQmlScriptStringPrivate
This allows constructing QQmlScriptStrings from C++, which we need for
moving the logic of the various text editing context menu action types
to C++ and keeping the QML focused on styling.

For example, QQuickCopyAction can now assign the following binding from
C++:

    editor.selectedText.length > 0

This will be auto-tested as part of a follow-up commit.

Task-number: QTBUG-133556
Task-number: QTBUG-134903
Change-Id: Ic291cdf86519194dd9c8a470b9cd8d4914f89859
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>