QQ4A: Add possibility to register signals before QML is loaded
This commit adds a queue mechanism and API in QAndroidViewSignalManager
that allows callers to add signal listeners before the QML view has
been loaded.

API Added to QAndroidViewSignalManager always returns true, but might
still fail if a signal with the desired signalName does not exist. If
using the QtQuickViewContent API, this should never happen, but with
the QtQuickView API, this is a possibility with typos or signal names
that are computed with some logic in code.

Internally, QAndroidViewSignalManager listens to the
QQuickView::statusChanged signal to detect when the QML view is done
loading.

Task-number: QTBUG-137057
Change-Id: I2ea95f096517f2b7f45a60f236923d42cb5cf1fe
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>