qmllint/Quick: Prevent attached property warning in custom parser
Using attached properties in PropertyChanges is not the most common use
case, but it is supported. It should thus not create any warnings.

In general, any custom parser can do whatever it desires with the
bindings it receives. Consequently we skip validation of attached
properties when inside a custom parser.

Task-number: QTBUG-138164
Pick-to: 6.10
Change-Id: I5a523518b6983adb96c2e1ccac7d7944c1956aa5
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>