Deprecate QQmlSslConfiguration::sslOptions property
The property has two major issues:
* it is out of sync with the underlying m_configuration
* it does not allow to disable an SSL option once it was enabled

Fixing the behavior is considered dangerous, because it affects the
default SSL options, and so potentially affects the application
security.

As a result, it was decided to deprecate this property and introduce
a new one instead of it.

Move the QML tests of the deprecated property into a separate TestCase
and enable it conditionally based on the value of
QT_REMOVAL_QT7_DEPRECATED_SINCE(6, 11) macro. This requires to use
QUICK_TEST_MAIN_WITH_SETUP() macro and introduce a custom Setup class
to pass the proper value from C++ to QML.

[ChangeLog][QQmlNetwork][Deprecation Notices] Deprecated the sslOptions
property of sslConfiguration/sslDtlsConfiguration. Use the new
sslOptionFlags property instead.

Fixes: QTBUG-137900
Change-Id: I7cd559e32475272fa77ae9dbf735d70371755725
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>