FluentWinUI3: Improve RadioButton when using high contrast themes
The RadioButton's indicator would not use the same colors as the WinUI3
RadioButton for users that had enabled a high contrast theme.

The indicator background is normally a ColorImage. But for high contrast
themes we prefer to keep things simple, and draw simple colors on top of
the image. It is possible to change the binding on the ColorImage to use
desired high contrast theme colors from the palette, but doing so will
slightly alter the colors used, due to blending with the pixel values
from the underlying png image used by ColorImage. Thus we put a
rectangle on top, so that high contrast themes become independent from
it.

Task-number: QTBUG-129088
Pick-to: 6.10
Change-Id: I57ef286584f7df2c26bacbc0bc9f3bcb7a1a64cd
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>