tst_qqmlpreview: remove dead code
This code was meant to add environment variable to a process before
starting it. However, the "executable" argument refers to the path of the qml tool binary itself, not the path to a qml file. Therefore, comparing it against the QML file path is incorrect and redundant. Instead, extend the function signature of the method where process is started.

Amends 64cd070c8c4bee26c04bf48dcd99a29262a2fd6f.

Change-Id: I6d5e40750fa9f483a6e21ede27980a0abd872d37
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>